move timeline utils functions to new file
